having module register error/exception handlers in Module.php
Fixes #1. Previously, this module was registering the error and
exception handlers in the sendException call, which is only ever called
when a `dispatch.error` is handled.  For most apps, this is probably
already too late to register exception/error handlers.  Plus, if an app
never encounters a `dispatch.error` event, the handlers will never be
registered, so any errors/exceptions that don’t trigger a
`dispatch.error` wouldn’t get sent to Bugsnag.

Moving the registration of these handlers to the onBootstrap method of
Module.php, which should get them registered pretty early in the app’s
lifecycle.

Tested that errors and exceptions that don’t trigger a `dispatch.error`
event are caught by the handlers, and sent to BugSnag.